requirements.txt
asgiref==3.7.2
celery==4.4.7
Django==3.2.1
django-cors-headers==4.1.0
djangorestframework==3.11.0
psycopg2==2.8.6
python-decouple==3.1
pytz==2023.3
requests==2.24.0
sqlparse==0.4.4
typing_extensions==4.6.3
